lyrics-fetcher/requirements.txt
# lyrics-fetcher/requirements.txt
pika==1.3.2
syncedlyrics==1.0.1
openai-whisper==20250625
torch==2.4.1
torchaudio==2.4.1
numpy==1.26.4
numba==0.60.0
rapidfuzz==3.10.1
ffmpeg-python==0.2.0
